+++ src/api/http/controller/customer_controller.go
@@ -12,6 +12,7 @@ import (
 
 var CustomerNotFound = errors.New("customer not found")
 var CustomerFieldShouldBeUnique = errors.New("customer field should be unique: ")
+var CustomerIdNotSet = errors.New("customer id should be set")
 
 type CustomerController struct {
 	Service *services.CustomerService
@@ -36,6 +37,9 @@ func (controller *CustomerController) List(c fiber.Ctx) error {
 
 func (controller *CustomerController) Get(c fiber.Ctx) error {
 	id := c.Params("id")
+	if id == "" {
+		return c.Status(fiber.StatusBadRequest).SendString(CustomerIdNotSet.Error())
+	}
 
 	res, err := controller.Service.GetCustomer(id)
 	if err != nil {
@@ -68,6 +72,9 @@ func (controller *CustomerController) Create(c fiber.Ctx) error {
 
 func (controller *CustomerController) Update(c fiber.Ctx) error {
 	id := c.Params("id")
+	if id == "" {
+		return c.Status(fiber.StatusBadRequest).SendString(CustomerIdNotSet.Error())
+	}
 
 	req := new(customer.UpdateCustomerRequest)
 	if err := c.Bind().Body(req); err != nil {
@@ -84,6 +91,10 @@ func (controller *CustomerController) Update(c fiber.Ctx) error {
 
 func (controller *CustomerController) Delete(c fiber.Ctx) error {
 	id := c.Params("id")
+	if id == "" {
+		return c.Status(fiber.StatusBadRequest).SendString(CustomerIdNotSet.Error())
+	}
+
 	err := controller.Service.DeleteCustomer(id)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
